apps/ai-service/requirement.txt
gradio
ultralytics
langchain-core
langchain-ollama
Pillow